ubigint to_bits toaltstack, simplify the syntax, cargo fmt

+++ src/ubigint/mod.rs
@@ -4,4 +4,4 @@ mod cmp;
 mod std;
 mod sub;
 
-pub struct UBigIntImpl<const N_BITS: usize>;
+pub struct UBigIntImpl<const N_BITS: u32>;
